Significant refactoring. Moved calcIdxs from MPDFcalculator class to MagSpecies and MagStructure classes. Renamed magIdxs as strucIdxs. Implemented more correct correlation length, where the magnitudes of the spins are changed directly rather than just multiplying by an exponential envelope. Moved all the helper functions into a new file magutils.py.
