+++ diffpy/mpdf/tests/testmpdffromcif.py
@@ -25,7 +25,7 @@ class mPDFfromCIFtest(unittest.TestCase):
         path = os.path.dirname(os.path.abspath(__file__))
         struc=loadStructure(find('MnO_cubic.cif',path))
         msp=diffpy.mpdf.MagSpecies(struc=struc)
-        msp.magIdxs=[0,1,2,3]
+        msp.strucIdxs=[0,1,2,3]
         msp.basisvecs=np.array([[1,-1,0]])
         msp.kvecs=np.array([[0.5,0.5,0.5]])
         msp.ffparamkey='Mn2'
@@ -35,7 +35,8 @@ class mPDFfromCIFtest(unittest.TestCase):
         mc=diffpy.mpdf.MPDFcalculator(magstruc=mstr)
         r,fr,dr=mc.calc(both=True)
         testval=np.round(dr[100],decimals=4)
-        self.assertEqual(testval,20.7885)
+        print(testval)
+        self.assertEqual(testval,20.7848)
 
 # End of class mPDFfromCIFtest
