+++ diffpy/mpdf/tests/testmpdffromselfcreated.py
@@ -17,15 +17,15 @@ class mPDFfromSelftest(unittest.TestCase):
         msp.spinBasis=np.array([[0,0,1],[0,0,-1]])
         mstr=diffpy.mpdf.MagStructure()
         mstr.loadSpecies(msp)
-        #mstr.makeAtoms()
-        #mstr.makeSpins()
+        mstr.makeAtoms()
+        mstr.makeSpins()
 
         # Set up the mPDF calculator
-        #mc=diffpy.mpdf.MPDFcalculator(magstruc=mstr)
-        #r,fr=mc.calc()
-        #testval=np.round(fr.max(),decimals=4)
-        #self.assertEqual(testval,17.0428)
-        self.assertEqual(1.0,1.0)
+        mc=diffpy.mpdf.MPDFcalculator(magstruc=mstr)
+        r,fr=mc.calc()
+        testval=np.round(fr.max(),decimals=4)
+        self.assertEqual(testval,17.0428)
+        #self.assertEqual(1.0,1.0)
 
 # End of class mPDFfromSelftest
